refactor: 코드 리팩토링 - 불필요한 패키지 디렉토리 줄임 - 서버관련 내용 Server 클래스로 구현
