--- wavnes/capture/__init__.py
@@ -1,1 +1,0 @@
-from .sniffer_manager import SnifferManager
