--- wavnes/capture/handlers/__init__.py
@@ -1,2 +1,0 @@
-from .packet_handler import PacketHandler
-from .mqtt_handler import MQTTHandler
