--- wavnes/capture/handlers/packet_handler.py
@@ -1,7 +1,0 @@
-from abc import ABC, abstractmethod
-
-
-class PacketHandler(ABC):
-    @abstractmethod
-    def process_packet(self, packet):
-        pass
