+++ wavnes/packet_handlers.py
@@ -1,0 +1,45 @@
+from abc import ABC, abstractmethod
+import time
+from scapy.all import *
+from scapy.contrib.mqtt import *
+
+
+def packet_time_info(start_time, previous_time, packet):
+    seconds_since_previous = float(packet.time - previous_time)
+    return {
+        'timestamp': '{:.6f}'.format(packet.time),
+        'time_of_day': time.strftime('%Y-%m-%d %H:%M:%S', time.localtime(packet.time)),
+        'seconds_since_beginning': '{:.6f}'.format(float(packet.time - start_time)),
+        'seconds_since_previous': '{:.6f}'.format(seconds_since_previous),
+    }
+
+
+class PacketHandler(ABC):
+    def __init__(self, packet):
+        self.packet_info = {
+            'source_ip': str(packet[IP].src),
+            'destination_ip': str(packet[IP].dst),
+            'source_port': int(packet[TCP].sport),
+            'destination_port': int(packet[TCP].dport),
+            'length': int(packet.len)
+        }
+
+    @abstractmethod
+    def process_packet(self, packet):
+        pass
+
+
+class MQTTHandler(PacketHandler):
+    def process_packet(self, packet):
+        mqtt_packet = packet[MQTT]
+
+        if hasattr(mqtt_packet, 'flags'):
+            self.packet_info['flags'] = int(mqtt_packet.flags)
+
+        if hasattr(mqtt_packet, 'topic'):
+            self.packet_info['topic'] = str(mqtt_packet.topic)
+
+        if hasattr(mqtt_packet, 'value'):
+            self.packet_info['value'] = mqtt_packet.value.decode('utf-8')
+
+        return self.packet_info
