--- wavnes/websocket/__init__.py
@@ -1,2 +1,0 @@
-from .websocket_handler import websocket_handler
-from .websocket_server import start_websocket_server
